requirements.txt
streamlit
plotly
pandas
pdfplumber
PyPDF2
python-docx
pdf2image
pytesseract
sentence-transformers
faiss-cpu
spacy
nltk
transformers
accelerate
safetensors
reportlab
numpy
google-generativeai
